ops: Split local and remote ops to separate files
