+++ test/compat_test.py
@@ -2,14 +2,14 @@
 import os
 
 from cola import compat
-from cola import operations
+from cola import operations_local
 
 
 def test_setenv():
     """Test the core.decode function"""
     key = 'COLA_UNICODE_TEST'
     value = '字龍'
-    ops = operations.LocalOperations()
+    ops = operations_local.LocalOperations()
     compat.setenv(ops, key, value)
     assert key in os.environ
     assert os.getenv(key)
