+++ test/helper.py
@@ -11,7 +11,7 @@ from cola import core
 from cola import git
 from cola import gitcfg
 from cola import gitcmds
-from cola import operations
+from cola import operations_local
 from cola.models import main
 
 # prevent unused imports lint errors.
@@ -91,7 +91,7 @@ def app_context():
 
     initialize_repo()
     context = Mock()
-    context.ops = operations.LocalOperations()
+    context.ops = operations_local.LocalOperations()
     context.git = git.create(ops=context.ops)
     context.git.set_worktree(core.getcwd())
     context.cfg = gitcfg.create(context)
